FDISK LITE implementation (no UI, 50K binary instead of 72K)
