+++ SOURCE/FDISK/USERINT2.C
@@ -16,38 +16,13 @@
 #include "main.h"
 #include "pcompute.h"
 #include "pdiskio.h"
+#include "userint0.h"
 #include "userint1.h"
 #include "userint2.h"
 #include "ansicon.h"
 
 #include "svarlang\svarlang.h"
 
-int IsRecognizedFatPartition( unsigned partitiontype )
-{
-   switch ( partitiontype ) {
-   case 1:
-   case 4:
-   case 6:
-      return TRUE;
-   case 0x0e:
-      if ( flags.version == W95 || flags.version == W95B ||
-           flags.version == W98 ) {
-         return TRUE;
-      }
-      break;
-   case 0x0b:
-      if ( flags.version == W95B || flags.version == W98 ) {
-         return TRUE;
-      }
-      break;
-   case 0x0c:
-      if ( flags.version == W95B || flags.version == W98 ) {
-         return TRUE;
-      }
-      break;
-   }
-   return FALSE;
-}
 
 /* Ask user if they want to use large disk support (FAT 32) */
 void Ask_User_About_FAT32_Support( void )
@@ -771,296 +746,6 @@ void Delete_Primary_DOS_Partition_Interface( void )
    }
 }
 
-/* Display information for all hard drives */
-void Display_All_Drives( void )
-{
-   int current_column_offset_of_general_drive_information;
-   int current_column_offset = 4;
-   int current_line = 3;
-   int current_line_of_general_drive_information;
-   int drive = 1;
-   int drive_letter_index = 0;
-   int index;
-
-   long space_used_on_drive_in_MB;
-
-   unsigned long usage;
-
-   Determine_Drive_Letters();
-
-   Print_At( 2, 2, "Disk   Drv   Mbytes    Free  Usage" );
-
-   for ( drive = 1; drive <= flags.maximum_drive_number - 127; drive++ ) {
-      if ( current_line > 18 ) {
-         current_line = 3;
-         current_column_offset = 45;
-
-         Print_At( 43, 2, "Disk   Drv   Mbytes    Free  Usage" );
-      }
-
-      /* Print physical drive information */
-      current_column_offset_of_general_drive_information =
-         current_column_offset;
-      current_line_of_general_drive_information = current_line;
-      space_used_on_drive_in_MB = 0;
-
-      /* Print drive number */
-      Position_Cursor( current_column_offset_of_general_drive_information,
-                       current_line );
-      Color_Printf( "%d", drive );
-
-      /* Print size of drive */
-
-      if ( !part_table[drive - 1].usable ) {
-         printf( "    -------- unusable ---------" );
-         current_line++;
-         continue;
-      }
-
-      Position_Cursor(
-         ( current_column_offset_of_general_drive_information + 10 ),
-         current_line );
-      Print_UL( part_table[drive - 1].disk_size_mb );
-
-      /* Get space_used_on_drive_in_MB */
-      for ( index = 0; index <= 3; index++ ) {
-         if ( ( part_table[drive - 1].pri_part[index].num_type != 5 ) &&
-              ( part_table[drive - 1].pri_part[index].num_type != 15 ) &&
-              ( part_table[drive - 1].pri_part[index].num_type != 0 ) ) {
-            space_used_on_drive_in_MB +=
-               part_table[drive - 1].pri_part[index].size_in_MB;
-         }
-      }
-
-      for ( index = 0; index < MAX_LOGICAL_DRIVES; index++ ) {
-         if ( part_table[drive - 1].log_drive[index].num_type > 0 ) {
-            space_used_on_drive_in_MB +=
-               part_table[drive - 1].log_drive[index].size_in_MB;
-         }
-      }
-
-      /* Print logical drives on disk, if applicable */
-
-      drive_letter_index = 0;
-      do {
-         if ( drive_lettering_buffer[drive - 1][drive_letter_index] > 0 ) {
-            current_line++;
-
-            if ( current_line > 18 ) {
-               current_line = 3;
-               current_column_offset = 45;
-
-               Print_At( 43, 2, "Disk   Drv   Mbytes   Free   Usage" );
-            }
-
-            /* Print drive letter of logical drive */
-            if ( ( ( drive_lettering_buffer[drive - 1][drive_letter_index] >=
-                     'C' ) &&
-                   ( drive_lettering_buffer[drive - 1][drive_letter_index] <=
-                     'Z' ) ) ||
-                 ( flags.del_non_dos_log_drives == TRUE ) ) {
-               Position_Cursor( ( current_column_offset + 6 ), current_line );
-               printf(
-                  "%c:",
-                  drive_lettering_buffer[drive - 1][drive_letter_index] );
-            }
-            else {
-               Position_Cursor( ( current_column_offset + 8 ), current_line );
-            }
-
-            /* Print size of logical drive */
-            Position_Cursor( ( current_column_offset + 10 ), current_line );
-
-            if ( drive_letter_index < 4 ) {
-               Print_UL( part_table[drive - 1]
-                            .pri_part[drive_letter_index]
-                            .size_in_MB );
-            }
-            else {
-
-               Print_UL( part_table[drive - 1]
-                            .log_drive[( drive_letter_index - 4 )]
-                            .size_in_MB );
-            }
-         }
-
-         drive_letter_index++;
-      } while ( drive_letter_index < 27 );
-
-      /* Print amount of free space on drive */
-      if ( part_table[drive - 1].disk_size_mb > space_used_on_drive_in_MB ) {
-         Position_Cursor(
-            ( current_column_offset_of_general_drive_information + 18 ),
-            current_line_of_general_drive_information );
-         Print_UL( part_table[drive - 1].disk_size_mb -
-                   space_used_on_drive_in_MB );
-      }
-
-      /* Print drive usage percentage */
-      if ( space_used_on_drive_in_MB == 0 ) {
-         usage = 0;
-      }
-      else {
-         usage = Convert_To_Percentage( space_used_on_drive_in_MB,
-                                        part_table[drive - 1].disk_size_mb );
-      }
-
-      Position_Cursor(
-         ( current_column_offset_of_general_drive_information + 28 ),
-         current_line_of_general_drive_information );
-      printf( "%3d%%", usage );
-
-      current_line++;
-   }
-
-   Print_At( 4, 20, "(1 Mbyte = 1048576 bytes)" );
-}
-
-void Display_CL_Partition_Table( void )
-{
-   int index = 0;
-
-   unsigned long usage = 0;
-   Partition_Table *pDrive = &part_table[flags.drive_number - 0x80];
-
-   Determine_Drive_Letters();
-
-   printf( "\nCurrent fixed disk drive: %1d",
-           ( flags.drive_number - 127 ) );
-
-   printf( "   %lu sectors, geometry %lu/%03lu/%02lu", pDrive->disk_size_sect,
-           pDrive->total_cyl + 1, pDrive->total_head + 1,
-           pDrive->total_sect );
-
-   if ( !Is_Pri_Tbl_Empty() ) {
-      printf( "\n\nPartition   Status   Mbytes   Description      Usage" );
-      printf( "    Start CHS       End CHS\n" );
-   }
-   else {
-      printf("\n\nNo partitions defined.\n");
-   }
-
-   index = 0;
-   do {
-      if ( pDrive->pri_part[index].num_type > 0 ) {
-         /* Drive Letter of Partition */
-         if ( IsRecognizedFatPartition( pDrive->pri_part[index].num_type ) ) {
-            printf(
-               " %1c:",
-               drive_lettering_buffer[( flags.drive_number - 128 )][index] );
-         }
-         else {
-            printf( "   " );
-         }
-
-         /* Partition Number */
-         printf( " %1d", ( index + 1 ) );
-
-         /* Partition Type */
-         printf( " %3d", ( pDrive->pri_part[index].num_type ) );
-
-         /* Status */
-         if ( pDrive->pri_part[index].active_status > 0 ) {
-            printf( "      A" );
-         }
-         else {
-            printf( "       " );
-         }
-
-         /* Mbytes */
-         printf( "    " );
-         Print_UL( pDrive->pri_part[index].size_in_MB );
-
-         /* Description */
-         printf( "   %-15s",
-                 partition_lookup_table_buffer_long[pDrive->pri_part[index]
-                                                       .num_type] );
-
-         /* Usage */
-         usage = Convert_To_Percentage( pDrive->pri_part[index].size_in_MB,
-                                        pDrive->disk_size_mb );
-
-         printf( "   %3lu%%", usage );
-
-         /* Starting Cylinder */
-         printf( "%6lu/%03lu/%02lu", pDrive->pri_part[index].start_cyl,
-                 pDrive->pri_part[index].start_head,
-                 pDrive->pri_part[index].start_sect );
-
-         /* Ending Cylinder */
-         printf( " %6lu/%03lu/%02lu", pDrive->pri_part[index].end_cyl,
-                 pDrive->pri_part[index].end_head,
-                 pDrive->pri_part[index].end_sect );
-         printf( "\n" );
-      }
-
-      index++;
-   } while ( index < 4 );
-   printf(
-      "\nLargest continious free space for primary partition = %lu MBytes\n",
-      Max_Pri_Free_Space_In_MB() );
-
-   /* Check to see if there are any drives to display */
-   if ( ( brief_partition_table[( flags.drive_number - 128 )][4] > 0 ) ||
-        ( brief_partition_table[( flags.drive_number - 128 )][5] > 0 ) ) {
-      printf( "\nContents of Extended DOS Partition:\n" );
-      printf( "Drv Volume Label  Mbytes  System   Usage" );
-      printf( "    Start CHS       End CHS\n" );
-
-      /* Display information for each Logical DOS Drive */
-      index = 4;
-      do {
-         if ( brief_partition_table[( flags.drive_number - 128 )][index] >
-              0 ) {
-            if ( IsRecognizedFatPartition( brief_partition_table[(
-                    flags.drive_number - 128 )][index] ) ) {
-               /* Display drive letter */
-               printf( " %1c:", drive_lettering_buffer[( flags.drive_number -
-                                                         128 )][index] );
-
-               /* Display volume label */
-               printf( " %11s", pDrive->log_drive[index - 4].vol_label );
-            }
-            else {
-               printf( "               " );
-            }
-
-            /* Display size in MB */
-            printf( "  " );
-            Print_UL( pDrive->log_drive[( index - 4 )].size_in_MB );
-
-            /* Display file system type */
-            printf( "  %-8s",
-                    partition_lookup_table_buffer_short
-                       [pDrive->log_drive[( index - 4 )].num_type] );
-
-            /* Display usage in % */
-            usage = Convert_To_Percentage(
-               pDrive->log_drive[index - 4].num_sect, pDrive->ext_num_sect );
-
-            printf( "  %3lu%%", usage );
-
-            /* Starting Cylinder */
-            printf( "%6lu/%03lu/%02lu",
-                    pDrive->log_drive[index - 4].start_cyl,
-                    pDrive->log_drive[index - 4].start_head,
-                    pDrive->log_drive[index - 4].start_sect );
-
-            /* Ending Cylinder */
-            printf( " %6lu/%03lu/%02lu", pDrive->log_drive[index - 4].end_cyl,
-                    pDrive->log_drive[index - 4].end_head,
-                    pDrive->log_drive[index - 4].end_sect );
-            printf( "\n" );
-         }
-
-         index++;
-      } while ( index < 27 );
-      printf(
-         "\nLargest continious free space in extended partition = %lu MBytes\n",
-         Max_Log_Free_Space_In_MB() );
-   }
-   printf("\n");
-}
 
 /* Display Extended Partition Information Sub Screen */
 void Display_Extended_Partition_Information_SS( void )
@@ -1457,19 +1142,6 @@ void Display_Primary_Partition_Information_SS( void )
    }
 
    printf( " Mbytes (1 Mbyte = 1048576 bytes)" );
-}
-
-/* Dump the partition tables from all drives to screen */
-void Dump_Partition_Information( void )
-{
-   int index = 0;
-   //flags.extended_options_flag=TRUE;
-
-   do {
-      flags.drive_number = index + 128;
-      Display_CL_Partition_Table();
-      index++;
-   } while ( index + 128 <= flags.maximum_drive_number );
 }
 
 /* List the Partition Types */
