+++ SOURCE/FDISK/USERINT2.H
@@ -13,20 +13,16 @@ void Change_Current_Fixed_Disk_Drive( void );
 void Delete_Extended_DOS_Partition_Interface( void );
 void Delete_N_DOS_Partition_Interface( void );
 void Delete_Primary_DOS_Partition_Interface( void );
-void Display_All_Drives( void );
-void Display_CL_Partition_Table( void );
 void Display_Extended_Partition_Information_SS( void );
 void Display_Help_Screen( void );
 void Display_Or_Modify_Logical_Drive_Information( void );
 void Display_Partition_Information( void );
 void Display_Primary_Partition_Information_SS( void );
-void Dump_Partition_Information( void );
 void Interactive_User_Interface( void );
 void List_Partition_Types( void );
 void Menu_Routine( void );
 void Modify_Extended_Partition_Information( int logical_drive_number );
 void Modify_Primary_Partition_Information( int logical_drive );
-void Pause( void );
 void Position_Cursor( int row, int column );
 void Print_Centered( int y, char *text, int style );
